Support Custom VMs
This changeset supports use of custom Graal VMs, by way of (1) a custom download URL template, and (2) custom fingerprints for the downloaded VM binary. This allows someone to invoke the `graal_bindist_repository` with the following pattern to use a custom VM:

```python
graal_bindist_repository(
  version = "19.3.1",
  java_version = "8",
  edition = "ce",
  url_template = "https://my-internal-server.local/graalvm-{platform}-{version}-{java_version}.tar.gz",
  fingerprints = {
    "8": {
        "darwin-amd64": "<some-sha-here>",
        "linux-amd64": "<some-sha-here>",
    },
    "11": {
        "darwin-amd64": "<some-sha-here>",
        "linux-amd64": "<some-sha-here>",
    },
  },
)
```

This will use the URL template, and hashes, in the invocation rather than using the defaults.