fix: remove `module_ctx.extension_metadata`
In the current Bzlmod setup, all repositories generated by the `graalvm`
extension may or may not be used by users and whether they should be
can't be predicted based on the given tags: Users that want to generate
toolchains need the `<name>_toolchains` repo, users that rely on SDK
files need the `<name>` repo.

For now, remove the use of `module_ctx.extension_metadata` so that no
warning is shown when users register toolchains in `<name>_toolchains`.
In the future, we should improve the way toolchains are registered with
Bzlmod by automatically registering a toolchain configured via tags.

Signed-off-by: Fabian Meumertzheim <fabian@meumertzhe.im>